Fixed issue LogOut for authentication modes 'access_control_ie' and 'username_password' is crashing Visual studio instance for VS19
Fixed issue LogOut' for authentication modes 'access_control_ie' and 'username_password' is crashing visual studio instance for VS19  and VS22 and added log statement and showing message if no vulnerabilities found on scan completion .

+++ CxActionShared/Helpers/LoginHelper.cs
@@ -459,13 +459,19 @@ namespace CxViewerAction.Helpers
         internal static void DoLogout()
         {
             Application.DoEvents();
+            LoginData login = LoginData.GetLoginDataInstance;
             Logger.Create().Info("Logging out, clearing authentication data.");
             OidcLoginData oidcLoginData = OidcLoginData.GetOidcLoginDataInstance();
             oidcLoginData.AccessToken = null;
             oidcLoginData.RefreshToken = null;
             oidcLoginData.AccessTokenExpiration = -1;
             _isLogged = false;
-            Cef.GetGlobalCookieManager().DeleteCookies("", "");
+            if (string.IsNullOrWhiteSpace(login.AuthenticationType) && (login.AuthenticationType == Constants.AuthenticationaType_DefaultValue))
+            {
+                loginCache = null;
+                Cef.GetGlobalCookieManager().DeleteCookies("", "");                
+            }
+            
         }
 
         /// <summary>
